Update patch set 2
Patch Set 2:

The new patch just fixes ARCH_HAS_BIGENDIAN to _HAVE_ (for consistency reasons). Brian's comment still stands.. We will get back to you as soon as possible.

"you shouldn't need the patches from the 0.9.9-dev branch, we are using 1.0.0 which should include them? or are these not in the upstream tree?"

Patch-set: 2
Label: Code-Review=0