Update patch set 2
Patch Set 2:

So is this just the code from 1.0.1 and 1.1.0 (CVS HEAD) found here http://cvs.openssl.org/chngview?cn=21708 ?

If so, I'd still format it as a patch in patches/ including updating the openssl.config and patches/README. It should be layed out as it is upstream so when we update to a version that includes it we can simply drop the patch. also, don't include the .s files, change the import_openssl.sh script to run the .pl as it does for arm.

Patch-set: 2