Update patch set 1
Patch Set 1:

Exactly and you can also include in the patch your changes to make sure the scripts are run to generate the asm and include it in the build. it's just really import that the import script work out of the box so people can perform openssl upgrades for security issues without worrying about losing other patches.

Patch-set: 1