Update patch set 2
Patch Set 2: (2 inline comments)

Sorry, I missed the new patchset.

So there, is a typo in openssl.confg to fix.

Finally, you do want to add and checkin the generated *-mips.s files, its just the need to be in the patch to, since at import time the generated ones would be lost otherwise. but they need to be checked in, because the build system does do the import step on every build.

Patch-set: 2
Label: Verified=0