Update patch set 4
Patch Set 4:

Just the minimal necessary new stuff should be in the patch, which means the .pl files. 

the ./import_openssl.sh will apply the patch to create the .pl files in the local directory, then run them to generate the .s files. 

everything needed to compile (.s) needs to be checked in, even if generated, since the build won't run the import. sorry I wasn't clearer about this before.

and yes, sorry, use whatever version is current in AOSP (1.0.0e). then you should see that the only diff in gettit is your new additions.

I did test applying this internally and it doesn't conflict with the newer versions, but I do want to get this in before trying to upgrade to 1.0.1 since a new dot release usually requires more rework of our patches.

Patch-set: 4