Update patch set 2
Patch Set 2:

This problem has not yet been reported to openssl.org.
I will do that Wednesday.

The problem dates from June 5 2011, when wrappers were added to the set-key methods for AES, RC4, and camellia. They did the renaming of the generic non-asm methods, and the asm methods for only some of the many architectures previously supported.  This broke the Mips build and perhaps others.

The forcing of o32 abi is probably not appropriate to send upstream.

Patch-set: 2