Rename assembly files with .S extensions.
The openssl files were generated with an .s extension,
which tells GCC that they normally don't need to be sent
to the pre-processor (i.e. they are passed directly to the
assembler).

Unfortunately, all these files _do_ need to be preprocessed,
which is why 'LOCAL_AS_CFLAGS := -x assembler-with-cpp' was
required in Crypto.mk.

As simpler way to solve the issue is simply to use an .S
extension when generating the assembly files. GCC knows
that these must be pre-processed first.

So the patch does:

  - Rename all .s files to .S
  - Remove the use of -x assembler-with-cpp from Crypto.mk
  - Modify import_openssl.sh to directly generate .S files
    (tested)

Context: This makes it easier to reuse the exact same sources
         for Chrome on Android. Its gyp build system doesn't
         have a feature comparable to LOCAL_AS_CFLAGS.

Change-Id: I708d9fbcf8d42b5c39a7d30df2b03ed79a3e62f0