Update patch set 1
Patch Set 1:

Technically, this should not affect the Android build at all, but it's a nice thing to have for the Chromium build.

More specifically, for the "linux_redux" build which is an x64 build of Chromium for Linux that uses OpenSSL instead of NSS. This is never used in production, nor official supported, but it helps checking that all the OpenSSL-specific stuff in the Chromium source tree that doesn't directly depend on the Android platform APIs work properly (i.e. it's considerably easier to run unit tests for it).

Patch-set: 1
Label: Code-Review=0