Fix x86_64 assembly file generation.
The "sha512-x86_64.pl" script actually needs the name of the output
file as its second parameter to determine whether to generate SHA-256
or SHA-512 routines.

This patch does the following:

- Fix import_openssl.sh to invoke the script properly
- Add the generation of sha256-x86_64.S as well.

Note that this patch is the result of running:

  ./import_openssl.sh import /path/to/openssl-1.0.1c.tar.gz

Which means that no other source files were impacted by the change.

Only needed for the Chromium x86_64 "linux_redux" build and the
SPDY host proxy program (flip_in_mem_edsm_server).

Change-Id: Ia40737f5952c7b156bd51844571e4f759910a6a1