Update patch set 2
Patch Set 2:

I'm not familiar with this part of OpenSSL, but I fear that this change is insufficiently motivated. Normally errors remain so that the location that called OpenSSL can inspect them. Also, why free the filename? It seems that the _free function does that. Is this really a missing call to the _free function?

I think a slightly expanded CL message and perhaps a code comment or two can clear all that up.

Patch-set: 2