Update patch set 1
Patch Set 1:

Note: I don't claim this is the right fix. Maybe a patch file is not properly applied, just wanted to raise your attention on this point and ask for your input.

For more context: I'm working on small modifications to import_openssl.sh to make the result easier to import into the Chromium port. As such, I prefer starting from a clean slate.

And in case you were wondering, the issues are really minor:

- Some files are encoded in ISO-8859-1 and need to be converted into UTF-8, otherwise the Chromium review tool will completely barf on them :(

- The mips assembly files should be renamed from .s to .S

- Ensure there is a consistent way to list all source files required for a given CPU-specific build that can be used by both builds. I'm thinking about generating some of the content of Crypto.mk and Ssl.mk during the import_openssl.sh step from a set of simpler config files.

Patch-set: 1
Label: Code-Review=0