Convert ISO-8859-1 files to UTF-8.
This patch modifies import_openssl.sh to convert all untarred OpenSSL
source files that are in ISO-8859-1 encoding into UTF-8.

The main reason for this is that the Chromium review tool doesn't support
anything else (i.e. "git cl upload" will barf with a mysterious Python
exception if the uploaded diff files aren't UTF-8).

This makes it easier to import the Android sources into the Chromium tree,
and should have no impact on the build products.

Change-Id: I43df753c41f5d9ed853a4252d7d05c5bbced98b4