Update patch set 4
Patch Set 4:

(1 comment)

you can run vogar with --verbose to see what is going on. the problem with the emulator is that it probably is waiting for /sdcard to be available. Did you do something like "mksdcard 1024M sdcard.img" and then add "-sdcard scard.img" to the emulator command line?

mips image should boot, enh has been doing it recently. 

as I mentioned harmony tests are now in the tree.

the path should have been ~dalvik-prebuild should now be /home/dalvik-prebuild to do gPrecise changes. note it is -, not _

Patch-set: 4