Update patch set 4
Patch Set 4:

so one one high level thing that is missing for me here is the rationale. More generated files (and more code to generate them) doesn't seem a feature superficially. Is this so you can more easily use the files in the Chrome build? That would some worth the extra code.

Patch-set: 4