Update patch set 4
Patch Set 4:

Yes, the rationale is to be able to use the exact same sources for the Chromium build. I've written an import_from_android.sh script that essentially automates the whole process, as in:

  1/ Grab a given revision of the AOSP sources for external/openssl.

  2/ Grab the official openssl tarball from the version
     in openssl.version

  3/ Apply a couple Chromium-specific patches when needed.

  4/ Re-run "import_openssl.ssh import /path/to/openssl.sh"

  5/ Generate a gyp file from the content of openssl.config
     exclusively. Said file is used to drive the Chromium
     build.

This is not meant to be added to the AOSP source tree (though I can do that if you want), but it avoids a tedious and error-prone process each time the sources change in AOSP
(it's easy to forget one compiler define) and we want to import them into Chromium.

Right now, Adam rubber-stamps every import that I perform because the changes are always non-trivial and manually comparing with the AOSP source tree is difficult.

I could understand if you don't want to see this, and I'm open to alternative suggestions. I only care about automating the process to avoid human errors that could impact the project in bad ways. E.g. I could write an Android.mk parser that understands the declarations in the openssl .mk files, but this is more work / more difficult to get right, and is prone to fail if you update them with new features of the Android build system.

Patch-set: 4