Ensure iterating over a lhash table prevents unwanted resizes.
This patch modifies the implementation of lhash tables to ensure
that deleting or inserting items during a lh_doall or lh_doall_arg
call will not unexpectedly resize the buckets array, leading to
undefined behaviour.

See http://crbug.com/298606 for an example.

This is done by adding a new field ('iteration_state') to the
lhash_st structure. Note that the field is placed at the end
of the structure to preserve the OpenSSL ABI [1]

This also removes the 'evil' changes to the 'down_load' field,
which were performed to prevent this. Unfortunately, this didn't
work due to an incorrect load-comparison computation is
lh_delete() [2]

[1] Since all _LHASH objects are heap-allocated through lh_new(),
    client code that accesses fields directly will not be
    impacted if their offsets do not change.

[2] The original code used the following expression to determine
    when to contract the table during a lh_delete() operation:

       down_load >= (num_items * LOAD_MULT / num_nodes)

    Unfortunately, due to truncation towards zero, this could
    trigger even when down_load was set to 0 in certain cases.
    See the comments for the LH_SHOULD_CONTRACT() macro for
    more explanations.

Change-Id: I9fa5b84c5d787d41fd520b63f7e641cf22fbb5b6