Update patch set 1
Patch Set 1:

Hi Enh,
We are waiting for the review result for the patch:
https://android-review.googlesource.com/#/c/83078/

Once that patch is approved, we could re-visit this patch as your suggestion.

This patch is not necessary for ARM if no multilib build requirement. But it's necessary for x86_64 even no multilib build requirement.

Patch-set: 1