Update patch set 1
Patch Set 1:

The basic problem here is how to get two differently-configured versions of the library to coexist with each other.

This way, the consumer of the static version is responsible for #include-ing the static config header before anything else, or risk an incorrect configuration.  The advantage of this is that even if they don't, as long as they don't include any headers that rely on unconfigured functionality, it will still work.  E.g. if fastbootd (as it currently is) forgot to include this, it would still work correctly.

An alternative would be having consumers add a compile flag (e.g. LOCAL_CFLAGS += -DOPENSSL_USE_STATIC_CONFIG), and switch header configs on that.

Additionally, we could build an identically-configured static version of the library if we built a static version of libdl, with the caveat that it will probably do some crazy stuff/not work/crash if anyone ever tries to use DSO functionality.

Patch-set: 1
Label: Verified=0