Update patch set 2
Patch Set 2: Code-Review-1

I found part of the problems but not all solutions.
(1) Some .c files like crypto/bio/b_dump.c does not compile on mac with -no-integrated-as.
(2) Some .s files like crypto/modes/asm/ghash-armv4.S does not compile without -no-integrated-as.

For (1) the filter-out method in this patch could be used but should be applied to LOCAL_CLANG_CFLAGS_<target> not LOCAL_CLANG_ASFLAGS_<target>, at least for the b_dump.c file. The filter-out method is not enough as the following is defined in Crypto.mk:
  -LOCAL_CFLAGS += -no-integrated-as
We need to fix Crypto.mk too. I'm not sure if the no-integrated-as in Crypto.mk is needed or not for other reasons.

For (2) we cannot filter-out -no-integrated-as for .S files, or LOCAL_CLANG_ASFLAGS. Usually not all .S files for all targets have problem, so in Crypto.mk, we should replace LOCAL_ASFLAGS += -no-integrated-as with something like LOCAL_CLANG_ASFLAGS_<target> += -no-integrated-as.

Patch-set: 2
Reviewer: Chih-hung Hsieh <1052399@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1