Update patch set 1
Patch Set 1:

> should we really be moving to external/boringssl?

I think that would be a pleasing cleanup in time. However, there are lots of references to external/openssl in Android. This change can be committed (and reverted!) atomically, in one place. I think we have to do this step first and can then make external/boringssl mirror external/openssl and change the references at leisure.

Patch-set: 1