testssl script didn't properly detect errors
When upgrading openssl to 1.0.2 there are some tests that rely on
$? of ssltest for pass/fail, but when ssltest is replaced with
adb shell then $? is always 0, so the test results are incorrect.

To get around this I added a new script adb_wrap.sh which generates
a script that runs the command and echos the $? from the device.
Then the host side can get the output of the script and parse it
for the actual return code of ssltest.

Change-Id: I285e1045e336f60d4b04b00ad5bb60f3085d130f