Add Android.bp and regen scripts
This supports building the host static libraries:
  libopenssl_crypto
  libopenssl_ssl
and host binary `openssl`. The libraries are only intended to be used by
the openssl binary and the host cpython _ssl module.

The android/regen.sh script can be used when updating openssl, and
generates a number of source files using the original perl scripts, and
some Android-*.gen.bp files with the information from the upstream
configuration system.

Bug: 173151817
Test: prebuilts/build-tools/build-prebuilts.sh on linux & mac
Change-Id: If43b551a51be62e0b3abe6f9bb0993a0f51547a5